Updated framework to React, rewrote front-end
After some testing with API Gateway/Cognito/Amplify, I realized a lot of the issues I was running into in terms of implementing these properly could be alleviated by doing a reset of sorts, and using the React framework instead. This commit implements the React framework; any AWS functionality is still not implemented into the website.